server/requirements.txt
Django>=4.0.4,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2-binary>=2.9.3,<3.0.0
drf-spectacular>=0.22.1,<0.23.0
